More example code
This is very porly documented, but you should now have an example of reading in a dataset (pushed direclty to github in data/RAW[...]), doing a bit of stuff, and then saving a "cleaned" copy, which is ignored from github since it is "derived" from this script.  That means that using the dataset on github, anybody can create the "TEST_RAW_CLEANED.csv" file, so there's no need to push it and take up space.