changed where the data was located
I wanted to keep the test data separate from some real files I'm about to push.  See how there's an all red file and an all green file?  This means that github essentially has 2 copies of this file now.  One that's been deleted and one that's pretty much unconnected to the history.  It's not necessarily a problem, but is harder to follow, so we try to avoid this as much as possible.  Or if we have to move/rename stuff, make that a separate commit from making changes to the code/file